+++ packages/web/src/routes/+layout.svelte
@@ -7,6 +7,12 @@
   import { appVersion } from '$lib/utils/app-version'
 
   import { setContext } from 'svelte'
+  import ToolSection from '$lib/components/layout/ToolSection.svelte'
+  import Margin from '$lib/components/spacing/Margin.svelte'
+  import {
+    PUBLIC_REPORT_BUG_URL,
+    PUBLIC_REPORT_REQUEST_URL,
+  } from '$env/static/public'
 
   setContext('i18n', i18n)
 
@@ -14,6 +20,18 @@
   function onFontLoad(this: HTMLLinkElement): void {
     this.media = 'all'
   }
+
+  const reportRequestLinkAttributes = {
+    href: PUBLIC_REPORT_REQUEST_URL,
+    target: '_blank',
+    rel: 'external noopener noreferrer',
+  } as const
+
+  const reportBugLinkAttributes = {
+    href: PUBLIC_REPORT_BUG_URL,
+    target: '_blank',
+    rel: 'external noopener noreferrer',
+  } as const
 </script>
 
 <svelte:head>
@@ -92,6 +110,35 @@
 
 <div class="font-monospace" data-testid="layout-root">
   <slot></slot>
+
+  <ToolSection
+    id="development-report"
+    class="container mx-auto my-4 w-100 text-center d-flex flex-column align-items-center"
+  >
+    <a class="d-block ms-1" {...reportRequestLinkAttributes}>
+      {$i18n.t('report.request', { ns: 'page/index' })}
+      <i class="bi bi-send"></i>
+    </a>
+
+    <Margin space={2} />
+
+    <a class="d-block ms-1" {...reportBugLinkAttributes}>
+      {$i18n.t('report.bug', { ns: 'page/index' })}
+      <i class="bi bi-send"></i>
+    </a>
+
+    <hr class="w-100" />
+
+    <a
+      class="d-block ms-1"
+      href="https://github.com/tooppoo/ac6_assemble_tool/releases"
+      target="_blank"
+      rel="external noopener noreferrer"
+    >
+      Release Notes
+      <i class="bi bi-journal-text"></i>
+    </a>
+  </ToolSection>
 
   <hr class="my-3" />
